Pass prepared statements in HTTP request body
HTTP request headers have a 64k limit, by default 4k in Jetty.
Prepared statements can be much longer than that, so if the header
"X-Presto-Prepared-Statement-In-Body" is specified, the server
assumes that the query has been sent as a Jackson-serialized QuerySubmission
and parses the request appropriately.

In order to preserve backwards compatibility for clients, it is still
possible to send a raw query string if the above header is not sent.

Only /v1/statement supports sending the query via a QuerySubmission object,
since /v1/query and /v1/execute don't support prepared statements.